+++ libraries/template-haskell/changelog.md
@@ -34,6 +34,9 @@
 
   * The argument to `TExpQ` can now be levity polymorphic.
 
+  * The types of `ConP` and `conP` have been changed to allow for an additional list
+    of type applications preceding the argument patterns.
+
 ## 2.16.0.0 *TBA*
 
   * Add support for tuple sections. (#15843) The type signatures of `TupE` and